lsp: basic completion for keywords and functions
